cpu: remove cad submodule
It was inefficient to track changes in CAD repository as submodule.
There is no intention to track binary file versions (for topologies
and schematics), so the files often pushed within the same commit
and this is potential source of instability. This patch removes
dependency on CAD submodule, schematics and topologies will be
kept as independent repo.

Signed-off-by: 1801BM1@gmail.com