am4: M4 processor initial commit
M4 processor is the Soviet replica of of military avionics
version of PDP-11/34. Supports EIS ans FIS instructions,
has no MMU, implemented over Am2900 bit-slice chipset with
56x1K horizontal microcode.

The initial version of HDL sources is presented, no test
and debug was performed. The fixes will be provided in the
coming commits.

The meta-disassembler tool is not fully functional yet.

Signed-off-by: 1801BM1 <1801BM1@gmail.com>