am4: introduce synchronous model
The reverse engineering of M4 processor is completed, the synchronous
model running on real FPGA is introduced.

The synchronous M4 model:

  - runs on top of real FPGAs
  - supports Wishbone compatible front-end bus
  - ported and tested on multiple development boards

Also bug with ena_ms frequency was fixed, the 1MHz ena_us was
divided by 1024 instead of 1000 resulting in 976Hz ena_ms strobe
frequency (was missed "else" keyword). The system timer frequency
was not affected.

Signed-off-by: 1801BM1 <1801BM1@gmail.com>