f11: demultiplex the microinstruction bus
In the original F11 processor chipset the microinstruction bus is
multiplexed, on the low master clock phase the microinstruction
opcode is provided by Control Chip and on the high master clock
phase the micro-operation status is translated on the bus by the
Data Chip (and optionally m[15] might be pulled low by MMU Chip).

The patch introduces dedicated lines for the status:

  - mo - early status, available on the low clock and valid
    before the rising edge of the master clock. This enables
    Control Chip to start prefetching ROM content on the rising
    edge

  - mc - latched micro-operation status, getting valid on
    the rising edge of the master clock, and remains valid during
    the whole master clock period

The F11 module itself and all chipset modules no longer depend
on status provided over multiplexed microinstruction bus - all
of them use the newly introduced status lines. This is a preparation
step to make bus fully demultiplexed and unidirectional and transfer
only microinstruction opcodes over this.

Signed-off-by: 1801BM1 <1801BM1@gmail.com>