vm2: fix interrupt acknowledge timeout vector
If there is bus timeout happens in interrupt acknowledgement cycle
the exception in HALT mode on vector 274 should be risen.

The synchronous and Wishbone VM2 models tried to handle regular
bus exception on vector 4 instead. This happened because of
IAKO cycle flag was not held in the moment the interrupt matrix
processed its inputs (on pli_ack strobe).

In addition to that the synchronous model saw multiple timeout
abort signals and fell into the double bus timeout exception
with vector 174.

The original asynchronous VM2 model is not impacted by this bug.
All the VM1 models also were checked and found as not impacted ones.

This fix:

  - holds the interrupt acknowledgement flag while timeout flags
    is active to allow interrupt matrix recognize the special bus
    timeout case

  - eliminates the multiple timeout abort requests

Reported-by: xolod79
Signed-off-by: 1801BM1 <1801BM1@gmail.com>