vm3: fix bus address multiplexer for data transactions
There are two kinds of bus transactions - data exchange and instruction
prefetch. The address multiplexer provided the wrong data address,
the prefetch address was provided if the data/prefetch cycle was
interleaved with internal register access (MMU/PSW).

The test program to catch:

	.title	vm3bug
	.asect
	. = 0
	jmp	@#start
	. = 24
	.word	start
	.word	340
start:	mov	#172342, R0
	mov	@R0, @#100
	clr	@R0
	nop
	nop
	halt
	halt

The writing to address @#100 actually occured to @#42 - the address of
the next instruction being prefetched.

Reported-by: forth32 user @ zx-pk.ru
Signed-off-by: 1801BM1 <1801BM1@gmail.com>