vm3: fix physical address wraparound in 18-bit mode
There is the factory test "CFKTHB0 PDP 11/34 MEM MGMT" failed
on test 26, @PC 25130 on the systems with more than 32KW of memory.

The test checks the physical address wraparound in 18-bit mode,
SR3 AS bit is set to zero. KPAR4 is assigned with 007777 value
and virtual address is 100100, the resulting physical address
should be 00000000, but it was 01000000 due to bug in the MMU
adder Verilog code. Sure, the original 1801BM3 processor has
no this bug, it was present in the models only.

Reported-by: forth32 user @ zx-pk.ru
Signed-off-by: 1801BM1 <1801BM1@gmail.com>